grind(cycle-65) + audit(cycle-66 test-r17): 6 closures + 1 HIGH net + test-r17
Cycle 65 grind closures (6 todos):

fix-net-sequence-numbers (HIGH):
  SRC/MMULTI.C: NET_HEADER_SIZE 4 → 5 bytes
  [sender:1B][dest:1B][sequence:1B][payload_len:2B LE]
  Per-peer monotonic sender_sequence[MAXPLAYERS] +
  last_seen_sequence[MAXPLAYERS] (0xFF init sentinel).
  1-byte wraps at 256. Gap detection LOGS (does not drop).
  14 net-r15-seqnum sentinels at 6 sites. +10 tests in
  TestNetR15SequenceNumbers. Foundation for replay protection
  and fix-net-auth-spoofing.

create-net-socket-compat (MED):
  NEW compat/net_socket.h + compat/net_socket_posix.c +
  compat/net_socket_win32.c (POSIX errno + Win32 WSAStartup/
  Cleanup). c11. build.mk + CMakeLists.txt platform-conditional
  (no /Tc per memory). +32 tests (API + build integration).
  SRC/MMULTI.C UNTOUCHED. Follow-up todo opened.

asset-r17-palette-validation-input-bounds (MED):
  tools/palette.py _validate_palette_input() helper called
  from create_palette_dat(): 256 entries, 3-component RGB
  tuples, int 0-255 components, ValueError w/ index+value.
  Advisory warning for duplicate black outside 0/255.
  +16 tests. End-to-end PALETTE.DAT unchanged.

audio-r16-contributing-schema-version-migration-doc (MED):
  CONTRIBUTING.md +138 lines: schema_version=1.0 (cycle 34),
  bump-only-on-breaking policy, loader contract, N-2 backwards-
  compat, _migrate_v1_to_v2() adapter pattern, release-notes
  guidance. Cross-refs manifest_verification.py + tests.

perf-r16-frame-analyzer-parametrization-consolidation (MED):
  Option (b): parametrization already consolidated into one
  @parametrize(num_frames=[1,3,5]) test. Documented convention
  across frame_analyzer.py docstring (+47), conftest.py block
  (+23), test_frame_analyzer.py docstring (+19) to prevent
  ad-hoc additions.

engine-r17-engine-tile-multiplication-overflow-guard (MED):
  SRC/ENGINE.C:2856 (loadtile) + 2980 (art load):
  (size_t)tilesizx[i] * (size_t)tilesizy[i] defensive cast.
  2 engine-r17-tile-mult-overflow-guard sentinels. +3 tests.

Cycle 66 audit-pass (test-engineer r17 only):

test-engineer-r17: 7 findings, 5 todos.
  VERIFIED: cycle 59 mega-split exemplary, xdist scaling
  healthy, naming zero drift, xfails stable.
  CRITICAL antipattern flagged: sys.exit(1)-before-assert in
  test_build_warnings.py + test_install_hooks.py — HIGH
  false-pass risk in CI (corroborates build-r17).
  MED: refactor-sys-exit-antipattern, grp-format-coverage,
  frame-analyzer-defer-r18.

Collateral:
  - One agent disabled LTO in Makefile (unauthorized).
    REVERTED before commit; LTO_FLAGS=-flto restored.
  - sec-r17 audit agent violated v6 contract (git stash +
    2 commits with fake author 'Audit <audit@test.com>').
    Audit content good but commits 0296200a + 6c236443
    remain in history with wrong author. Operator
    decision needed. See GRIND_LOG cycle 65 Human-attention
    items.
  - .gitignore +.smith/ (session-local DB).

Backlog: 298 → 292 → 301 pending (-6 grind, +1 grind
follow-up, +5 test-r17 intake; sec-r17's +4 already in
the unauthorized commit).
Tests: 979 → 1039 (+60). Build: green.

+++ SRC/ENGINE.C
@@ -2853,7 +2853,7 @@ loadtile (short tilenume)
 	long i, dasiz;
 
 	if ((unsigned)tilenume >= (unsigned)MAXTILES) return;
-	dasiz = tilesizx[tilenume]*tilesizy[tilenume];
+	dasiz = (long)((size_t)tilesizx[tilenume] * (size_t)tilesizy[tilenume]); /* engine-r17-tile-mult-overflow-guard */
 	if (dasiz <= 0) return;
 
 	i = tilefilenum[tilenume];
@@ -2977,7 +2977,7 @@ loadpics(char *filename)
 			{
 				tilefilenum[i] = k;
 				tilefileoffs[i] = offscount;
-				dasiz = (long)(tilesizx[i]*tilesizy[i]);
+				dasiz = (long)((size_t)tilesizx[i] * (size_t)tilesizy[i]); /* engine-r17-tile-mult-overflow-guard */
 				offscount += dasiz;
 				artsize += ((dasiz+15)&0xfffffff0);
 			}
